Fixed error for packages in bundles
Problem: Packages like tabularx or array, Ctan returns a download link to download a collection of packages.
Previously, user could download one pkg of a collection, which would download whole collection, and then another pkg of that collection, which would download that whole collection again.

This is not possible anymore because I check if any installed pkg has same path on ctan